Title: 합과 차의 나눗셈, Time: 121ms, Memory: 24MB, Status: Wrong Answer - CodeTree

+++ 240312/합과 차의 나눗셈/divide-of-sum-and-sub.py
@@ -1,4 +1,4 @@
 a, b = map(float, input().split())
-c = (a + b) / (a - b)
+c = float((a + b) / (a - b))
 
 print(f"{round(c, 2)}")
